% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasXpress.R
\docType{package}
\name{canvasXpress-package}
\alias{canvasXpress}
\alias{canvasXpress-package}
\title{CanvasXpress Visualization Package}
\description{
A package to assist in creating visualizations in CanvasXpress in R.
}
\details{
CanvasXpress is a standalone JavaScript library for reproducible research
with complete tracking of data and end-user modifications stored in a single 
PNG image that can be played back.
}
\section{Function Overview}{

TBD
}

\section{More Information}{

\url{http://canvasxpress.org}
}

