% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasxpress_as_list.R
\name{canvasXpress.add.to.last}
\alias{canvasXpress.add.to.last}
\title{Converts a list in args to json that can be used by CanvasXpress.}
\usage{
canvasXpress.add.to.last(geom, ...)
}
\description{
Converts a list in args to json that can be used by CanvasXpress.
}
