% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasxpress_as_list.R
\name{canvasXpress.as.list}
\alias{canvasXpress.as.list}
\title{Converts a canvasXpress object to a list that can be used by CanvasXpress.}
\usage{
canvasXpress.as.list(data, x, y, z, ...)
}
\description{
Converts a canvasXpress object to a list that can be used by CanvasXpress.
}
