% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasxpress_as_list.R
\name{canvasXpress.bin.2d}
\alias{canvasXpress.bin.2d}
\title{Converts a list in args to json that can be used by CanvasXpress.}
\usage{
canvasXpress.bin.2d(...)
}
\description{
Converts a list in args to json that can be used by CanvasXpress.
}
