% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasxpress_as_list.R
\name{canvasXpress.density.2d.filled}
\alias{canvasXpress.density.2d.filled}
\title{Converts a list in args to json that can be used by CanvasXpress.}
\usage{
canvasXpress.density.2d.filled(...)
}
\description{
Converts a list in args to json that can be used by CanvasXpress.
}
