% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasXpress.R
\name{canvasXpress.ggplot}
\alias{canvasXpress.ggplot}
\title{HTML Widget Creation using GGPLOT input}
\usage{
canvasXpress.ggplot(
  ggplot_obj,
  pretty = FALSE,
  digits = 4,
  width = 600,
  height = 400,
  destroy = FALSE
)
}
\arguments{
\item{ggplot_obj}{ggplot object}

\item{pretty}{print tagged code (JSON/HTML) nicely - default = FALSE}

\item{digits}{display digits - default = 4}

\item{width}{plot width (valid CSS units) - default = 600px}

\item{height}{plot height (valid CSS units) - default = 400px}

\item{destroy}{used to indicate removal of a plot - default = FALSE}
}
\value{
htmlwidgets object
}
\description{
**NOTE:** This is in-development as of 1.38.0.9001
}
\section{More Information}{

\url{https://www.canvasxpress.org}
}

\examples{
\dontrun{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  library(ggplot2)
  my_plot <- ggplot(diamonds, aes(x=carat, y=price, color=cut)) + geom_point()
  canvasXpress.ggplot(my_plot)
}
}

}
