% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasxpress_as_list.R
\name{canvasXpress.hex}
\alias{canvasXpress.hex}
\title{Converts a list in args to json that can be used by CanvasXpress.}
\usage{
canvasXpress.hex(...)
}
\description{
Converts a list in args to json that can be used by CanvasXpress.
}
