% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasXpress.R
\docType{data}
\name{canvasXpress.last}
\alias{canvasXpress.last}
\title{Last canvasXpress object}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
canvasXpress.last
}
\description{
Last canvasXpress object
}
\keyword{datasets}
