% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasxpress_as_list.R
\name{canvasXpress.point}
\alias{canvasXpress.point}
\title{Converts a list in args to json that can be used by CanvasXpress.}
\usage{
canvasXpress.point(...)
}
\description{
Converts a list in args to json that can be used by CanvasXpress.
}
