% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasxpress_as_list.R
\name{canvasXpress.scale.color.manual}
\alias{canvasXpress.scale.color.manual}
\title{Converts a list in args to json that can be used by CanvasXpress.}
\usage{
canvasXpress.scale.color.manual(...)
}
\description{
Converts a list in args to json that can be used by CanvasXpress.
}
