% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasXpress.R
\name{canvasXpressOutput}
\alias{canvasXpressOutput}
\title{canvasXpressOutput}
\usage{
canvasXpressOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{tbd}

\item{width}{tbd}

\item{height}{tbd}
}
\value{
Output function that enables the use of the widget in applications
}
\description{
Output creation function for canvasXpressOutput in Shiny applications and 
interactive Rmd documents
}
\seealso{
\link[canvasXpress]{renderCanvasXpress}
}

