% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_functionality.R
\name{cxHtmlPage}
\alias{cxHtmlPage}
\title{cxHtmlPage - create a stand-alone HTML page with embedded canvasXpress widget}
\usage{
cxHtmlPage(chartObject, width = "100\%", height = "70vh")
}
\value{
Stand-alone HTML page with embedded canvasXpress chart
}
\description{
#TODO
}
