% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_as_list.R
\name{ggplot.as.list}
\alias{ggplot.as.list}
\title{Converts a ggplot object to a list that can be used by CanvasXpress.}
\usage{
ggplot.as.list(o, ...)
}
\arguments{
\item{o}{the ggplot object}

\item{...}{additional parameters to the function}
}
\description{
Converts a ggplot object to a list that can be used by CanvasXpress.
}
