% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxPlot.R
\name{print.cxplot}
\alias{print.cxplot}
\title{Explicitly creates htmlwidgets object for output}
\usage{
\method{print}{cxplot}(x, ...)
}
\arguments{
\item{x}{plot to display}
}
\value{
htmlwidgets object
}
\description{
Explicitly creates htmlwidgets object for output
}

