% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasXpress.R
\name{renderCanvasXpress}
\alias{renderCanvasXpress}
\title{renderCanvasXpress}
\usage{
renderCanvasXpress(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{tbd}

\item{env}{tbd}

\item{quoted}{tbd}
}
\value{
Render function that enables the use of the widget in applications
}
\description{
Render function for canvasXpressOutput in Shiny applications and 
interactive Rmd documents
}
\seealso{
\link[canvasXpress]{canvasXpressOutput}
}

