% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasXpress.R
\docType{package}
\name{canvasXpress-package}
\alias{canvasXpress-package}
\alias{_PACKAGE}
\title{CanvasXpress Visualization Package}
\description{
A package to assist in creating visualizations in CanvasXpress in R.
}
\details{
CanvasXpress is a standalone JavaScript library for reproducible research
with complete tracking of data and end-user modifications stored in a single
PNG image that can be played back for an extensive set of visualizations.
}
\section{More Information}{

\url{https://canvasxpress.org}

\code{browseVignettes(package = "canvasXpress")}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/neuhausi/canvasXpress.git}
  \item Report bugs at \url{https://github.com/neuhausi/canvasXpress/issues}
}

}
\author{
\strong{Maintainer}: Connie Brett \email{connie@aggregate-genius.com}

Authors:
\itemize{
  \item Isaac Neuhaus \email{imnphd@gmail.com}
}

}
