% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasXpress.R
\docType{package}
\name{canvasXpress}
\alias{canvasXpress}
\alias{canvasXpress-package}
\title{CanvasXpress Visualization Package}
\usage{
canvasXpress(
  data = NULL,
  smpAnnot = NULL,
  varAnnot = NULL,
  graphType = "Scatter2D",
  events = NULL,
  afterRender = NULL,
  pretty = FALSE,
  digits = 4,
  width = 600,
  height = 400,
  destroy = FALSE,
  x = NULL,
  y = NULL,
  z = NULL,
  ...
)
}
\arguments{
\item{data}{data.frame-, matrix-, or list- classed data object}

\item{smpAnnot}{additional data that applies to samples (columns)}

\item{varAnnot}{additional data that applies to variables (rows)}

\item{graphType}{type of graph to be plotted - default = "Scatter2D"}

\item{events}{user-defined events (e.g. mousemove, mouseout, click and dblclick)}

\item{afterRender}{event triggered after rendering}

\item{pretty}{print tagged code (JSON/HTML) nicely - default = FALSE}

\item{digits}{display digits - default = 4}

\item{width}{plot width (valid CSS units) - default = 600px}

\item{height}{plot height (valid CSS units) - default = 400px}

\item{destroy}{used to indicate removal of a plot - default = FALSE}

\item{x}{used to designate the x object data value for cxplot functionality}

\item{y}{used to designate the y object data value for cxplot functionality}

\item{z}{used to designate the z object data value for cxplot functionality}

\item{...}{additional parameters passed to canvasXpress}
}
\value{
htmlwidgets object
}
\description{
A package to assist in creating visualizations in CanvasXpress in R.

Custom HTML widget creation function based on widget YAML and JavaScript for
use in any html-compatible context
}
\details{
CanvasXpress is a standalone JavaScript library for reproducible research
with complete tracking of data and end-user modifications stored in a single
PNG image that can be played back for an extensive set of visualizations.
}
\section{More Information}{

\url{https://www.canvasxpress.org}

\code{browseVignettes(package = "canvasXpress")}
}

