% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasXpress.R
\name{canvasXpress}
\alias{canvasXpress}
\title{canvasXpress}
\usage{
canvasXpress(data = NULL, smpAnnot = NULL, varAnnot = NULL,
  graphType = "Scatter2D", events = NULL, afterRender = NULL,
  pretty = FALSE, digits = 4, width = 600, height = 400, ...)
}
\value{
htmlwidget object
}
\description{
Custom HTML widget creation function based on widget YAML and JavaScript for 
use in any html-compatible context
}
