% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasXpress.R
\docType{package}
\name{canvasXpress}
\alias{canvasXpress}
\alias{canvasXpress-package}
\title{CanvasXpress Visualization Package}
\description{
A package to assist in creating visualizations in CanvasXpress in R
}
\section{Function Overview}{

TBD
}

\section{More Information}{

TBD
}

