% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasXpress.R
\docType{method}
\name{canvasXpress}
\alias{canvasXpress}
\title{canvasXpress}
\usage{
canvasXpress(data = NULL, decorData = NULL, smpAnnot = NULL,
  varAnnot = NULL, nodeData = NULL, edgeData = NULL, vennData = NULL,
  vennLegend = NULL, genomeData = NULL, graphType = "Scatter2D",
  events = NULL, afterRender = NULL, width = 600, height = 400,
  pretty = FALSE, digits = 4, ...)
}
\arguments{
\item{data}{tbd}

\item{decorData}{tbd}

\item{smpAnnot}{tbd}

\item{varAnnot}{tbd}

\item{nodeData}{tbd}

\item{edgeData}{tbd}

\item{vennData}{tbd}

\item{vennLegend}{tbd}

\item{genomeData}{tbd}

\item{graphType}{tbd}

\item{events}{tbd}

\item{afterRender}{tbd}

\item{width}{tbd}

\item{height}{tbd}

\item{pretty}{tbd}

\item{digits}{tbd}

\item{...}{tbd}
}
\value{
htmlwidget object
}
\description{
Custom HTML widget creation function based on widget YAML and JavaScript for 
use in any html-compatible context
}
\seealso{
\link[canvasXpress]{canvasXpressOutput}
}

