% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasXpress.R
\name{canvasXpress}
\alias{canvasXpress}
\title{canvasXpress}
\usage{
canvasXpress(data = NULL, decorData = NULL, smpAnnot = NULL,
  varAnnot = NULL, nodeData = NULL, edgeData = NULL, vennData = NULL,
  vennLegend = NULL, genomeData = NULL, newickData = NULL,
  graphType = "Scatter2D", events = NULL, afterRender = NULL,
  width = 600, height = 400, pretty = FALSE, digits = 4, ...,
  boxplotGroupData = NULL)
}
\arguments{
\item{data}{canvasXpress data frame
\emph{(rows are refered to variables; columns are refered to samples)}}

\item{decorData}{decoration/annotation data for the chart}

\item{smpAnnot}{additional data that applies to samples (columns)}

\item{varAnnot}{additional data that applies to variables (rows)}

\item{nodeData}{network visualization data nodes}

\item{edgeData}{network visualization data edges}

\item{vennData}{venn visualization data}

\item{vennLegend}{venn visualization legend}

\item{genomeData}{genome visualization data}

\item{newickData}{dendrogram data in Newick format}

\item{graphType}{type of graph to be plotted - default = 'Scatter2D'}

\item{events}{user-defined events (eg. mousemove, mouseout, click and dblclick)}

\item{afterRender}{event triggered after rendering}

\item{width}{plot width (valid CSS units) - default = 600px}

\item{height}{plot height (valid CSS units) - default = 400px}

\item{pretty}{print tagged code (json/html) nicely - default = FALSE}

\item{digits}{display digits - default = 4}

\item{...}{additional configuration options passed to canvasXpress}

\item{boxplotGroupData}{character label for grouped boxplot data.  If there is a 
non-null value and graphType is 'Boxplot' pre-calculated boxplot data is used from
the variables (iqr1, iqr3, qtl1, qtl3, median)}
}
\value{
htmlwidget object
}
\description{
Custom HTML widget creation function based on widget YAML and JavaScript for 
use in any html-compatible context
}

