% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasXpress.R
\name{canvasXpress}
\alias{canvasXpress}
\title{canvasXpress}
\usage{
canvasXpress(data = NULL, smpAnnot = NULL, varAnnot = NULL,
  graphType = "Scatter2D", events = NULL, afterRender = NULL,
  pretty = FALSE, digits = 4, width = 600, height = 400,
  destroy = FALSE, ...)
}
\arguments{
\item{data}{data.frame-, matrix-, or list- classed data object}

\item{smpAnnot}{additional data that applies to samples (columns)}

\item{varAnnot}{additional data that applies to variables (rows)}

\item{graphType}{type of graph to be plotted - default = "Scatter2D"}

\item{events}{user-defined events (e.g. mousemove, mouseout, click and dblclick)}

\item{afterRender}{event triggered after rendering}

\item{pretty}{print tagged code (JSON/HTML) nicely - default = FALSE}

\item{digits}{display digits - default = 4}

\item{width}{plot width (valid CSS units) - default = 600px}

\item{height}{plot height (valid CSS units) - default = 400px}

\item{destroy}{used to indicate removal of a plot - default = FALSE}

\item{...}{additional parameters passed to canvasXpress}
}
\value{
htmlwidgets object
}
\description{
Custom HTML widget creation function based on widget YAML and JavaScript for
use in any html-compatible context
}
