% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasXpress.R
\name{canvasXpressOutput}
\alias{canvasXpressOutput}
\title{canvasXpressOutput}
\usage{
canvasXpressOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{shiny unique ID}

\item{width}{width of the element - default = 100\%}

\item{height}{height of the element - default = 400px}
}
\value{
Output function that enables the use of the widget in applications
}
\description{
Output creation function for canvasXpressOutput in Shiny applications and 
interactive Rmd documents
}
\seealso{
\link[canvasXpress]{renderCanvasXpress}

\link[canvasXpress]{cxShinyExample}
}

