% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_examples.R
\name{cxShinyExample}
\alias{cxShinyExample}
\title{Create Shiny Example Application}
\usage{
cxShinyExample(example = NULL)
}
\arguments{
\item{example}{character name of a valid example application.}
}
\value{
Launches a running shiny example application
}
\description{
This function runs one of the available shiny example applications.  To see the 
list of available example applications run the function with a missing example 
value
}
\seealso{
\link[canvasXpress]{canvasXpressOutput}

\link[canvasXpress]{renderCanvasXpress}
}

