% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functionality.R
\name{renderCanvasXpress}
\alias{renderCanvasXpress}
\title{Shiny Render function}
\usage{
renderCanvasXpress(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{expression used to render the canvasXpressOutput}

\item{env}{environment to use - default = parent.frame()}

\item{quoted}{whether the expression is quoted - default = FALSE}
}
\value{
Render function that enables the use of the widget in applications
}
\description{
Render function for canvasXpressOutput in Shiny applications and
interactive Rmd documents
}
\section{Destroy}{

When there exists a need to visually remove a plot from a Shiny
application when it is not being immediately replaced with a new plot use
the destroy option as in:

\code{renderCanvasXpress({canvasXpress(destroy = TRUE)})}
}

\seealso{
\link[canvasXpress]{canvasXpressOutput}

\link[canvasXpress]{cxShinyExample}
}
